refactor: practice short 로직 hook으로 빼기, context랑 분리
