+++ client/src/components/typing/short/_hook/contextShortTyping.tsx
@@ -1,15 +1,11 @@
-import { useRouter } from 'next/router';
 import type { ReactNode } from 'react';
-import { useCallback } from 'react';
-import { createContext, useContext, useMemo, useRef, useState } from 'react';
+import { createContext, useContext } from 'react';
 
 import type { ShortTypingType } from '@/apis/typing';
 import Confirm from '@/components/common/Confirm';
-import ResultModal from '@/components/common/ResultModal/practice-mode';
-import useCurrentTyping from '@/components/typing/short/_hook/useCurrentTyping';
-import useToggle from '@/hooks/useToggle';
-import type { TypingHistoryType, TypingResultType } from '@/types/typing';
-import { getTypingHistoryAverage } from '@/utils/typing';
+import PracticeResultModal from '@/components/common/ResultModal/practice-mode';
+
+import usePracticeShortTyping from './usePracticeShortTyping';
 
 interface ContextShortTypingType {
   originalTyping: string;
@@ -25,14 +21,14 @@ interface ContextShortTypingType {
   prevOriginalTyping: string;
   nextOriginalTyping: string;
 
-  typingAvgResult: TypingResultType;
+  // typingAvgResult: TypingResultType;
 }
 
 interface ContextShortTypingHandlerType {
   onEndTyping: (input: string) => Promise<void>;
   onTyping: (inputChar: string) => void;
 
-  handleExitModalOpen: () => void;
+  onExit: () => void;
 }
 
 const ContextShortTyping = createContext<ContextShortTypingType | null>(null);
@@ -41,125 +37,20 @@ const ContextShortTypingHandler = createContext<ContextShortTypingHandlerType | 
 interface ShortTypingProviderProps {
   children: ReactNode;
   originalTypings: ShortTypingType[];
+  mode: 'PRACTICE' | 'test';
 }
 
-const ShortTypingProvider = ({ children, originalTypings }: ShortTypingProviderProps) => {
-  const router = useRouter();
-  const [currentIdx, setCurrentIdx] = useState(0);
-
+const ShortTypingProvider = ({ children, originalTypings, mode }: ShortTypingProviderProps) => {
   const {
-    originalTyping,
-    userTyping,
-    time,
-    typingCount,
-    typingAccuracy,
-    typingWpm,
-    typingSpeed,
-    handleTypingSubmit,
-    handleTyping,
-    timePlay,
-    timePause,
-  } = useCurrentTyping(
-    originalTypings[currentIdx] ?? {
-      typingId: 0,
-      content: '',
-    },
-  );
+    values: practiceValues,
+    actions: practiceActions,
+    endGameValue: practiceEndGameValue,
+  } = usePracticeShortTyping(originalTypings);
 
-  const [isResultModalOpen, handleResultModalToggle] = useToggle();
-  const [isExitModalOpen, handleExitModalToggle] = useToggle();
-
-  const history = useRef<TypingHistoryType[]>([]);
-  const prevUserTyping = history.current[history.current.length - 1]?.content ?? '';
-
-  const typingAvgResult = getTypingHistoryAverage(history.current);
-
-  const saveTypingHistory = useCallback(
-    (content: string) => {
-      history.current = [
-        ...history.current,
-        { typingSpeed, typingAccuracy, typingWpm, typingTime: time, content, endTime: new Date() },
-      ];
-    },
-    [time, typingAccuracy, typingSpeed, typingWpm],
-  );
-
-  const prevOriginalTyping = useMemo(
-    () => (currentIdx > 0 ? originalTypings[currentIdx - 1]?.content : ''),
-    [currentIdx, originalTypings],
-  );
-
-  const nextOriginalTyping = useMemo(
-    () => (currentIdx < originalTypings.length - 1 ? originalTypings[currentIdx + 1].content : ''),
-    [currentIdx, originalTypings],
-  );
-
-  const handleSubmit = useCallback(
-    async (input: string) => {
-      await handleTypingSubmit(input);
-      saveTypingHistory(input);
-    },
-    [handleTypingSubmit, saveTypingHistory],
-  );
-
-  const handleEndTyping = useCallback(
-    async (input: string) => {
-      // NOTE: 짧은글에서는 데이터를 전송하는 것을 기다리지 않고, 바로 다음 문장으로 넘어가는 것이 좋을 것 같다.
-      await handleSubmit(input);
-
-      if (currentIdx < originalTypings.length - 1) {
-        setCurrentIdx((prev) => prev + 1);
-      } else {
-        // 여기서 결과 모달을 띄워야하기 때문에, 모달 관련 로직을 여기서 처리하는게 맞을 것 같음
-        handleResultModalToggle();
-      }
-    },
-    [currentIdx, handleResultModalToggle, handleSubmit, originalTypings.length],
-  );
-
-  // NOTE: 모달 관련된 로직들을 분리할 방법이 있을까요?
-  // timePause, timePlay를 사용해야 하기떄문에 이곳에 만들었습니다.
-  const handleExitModalOpen = useCallback(() => {
-    timePause();
-    handleExitModalToggle();
-  }, [handleExitModalToggle, timePause]);
-
-  const handleExitModalClose = useCallback(() => {
-    timePlay();
-    handleExitModalToggle();
-  }, [handleExitModalToggle, timePlay]);
-
-  const handleResultModalOpen = useCallback(() => {
-    handleExitModalToggle();
-    handleResultModalToggle();
-  }, [handleExitModalToggle, handleResultModalToggle]);
-
-  const handleReplay = () => {
-    // NOTE : 다시하기 기능은 새로고침으로 구현했는데, 긴글에서도 새로고침으로 해도 괜찮을까요?
-    handleResultModalToggle();
-    router.reload();
-  };
-
-  // NOTE: 객체의 크기가 커서, 알아보기 힘들다면 쪼개는 것도 좋을 것 같다.
-  const values = {
-    originalTyping,
-    userTyping,
-    time,
-    typingCount,
-    typingWpm,
-    typingAccuracy,
-    typingSpeed,
-    prevUserTyping,
-    prevOriginalTyping,
-    nextOriginalTyping,
-    typingAvgResult,
-  };
-
-  const actions = {
-    onEndTyping: handleEndTyping,
-    onTyping: handleTyping,
-    handleExitModalOpen,
-  };
+  const values = mode === 'PRACTICE' ? practiceValues : practiceValues;
+  const actions = mode === 'PRACTICE' ? practiceActions : practiceActions;
+  const result = mode === 'PRACTICE' ? practiceEndGameValue.typingAvgResult : practiceEndGameValue.typingAvgResult;
+  const endGameValue = mode === 'PRACTICE' ? practiceEndGameValue : practiceEndGameValue;
 
   return (
     <ContextShortTyping.Provider value={values}>
@@ -168,19 +59,20 @@ const ShortTypingProvider = ({ children, originalTypings }: ShortTypingProviderP
 
         <Confirm
           header={'정말로 그만 두시겠어요?'}
-          isOpen={isExitModalOpen}
-          onClose={handleExitModalClose}
-          onAction={handleResultModalOpen}
+          isOpen={endGameValue.isExitModalOpen}
+          onClose={endGameValue.handleExitModalClose}
+          onAction={endGameValue.handleResultModalOpen}
           actionLabel='그만하기'
           closeLabel='계속하기'
         />
-        <ResultModal
-          isOpen={isResultModalOpen}
-          result={typingAvgResult}
-          endTime={history.current[history.current.length - 1]?.endTime ?? new Date()}
-          onAction={handleReplay}
-          actionLabel='다시하기'
-        />
+        {mode === 'PRACTICE' && (
+          <PracticeResultModal
+            isOpen={endGameValue.isResultModalOpen}
+            result={result}
+            endTime={endGameValue.endTime}
+            onReplay={endGameValue.handleReplay}
+          />
+        )}
       </ContextShortTypingHandler.Provider>
     </ContextShortTyping.Provider>
   );
